refactor did logic into a service - [CORE-1026]
