+++ dapps/profile.vue/src/components/did/DID.vue
@@ -67,6 +67,7 @@
       </div>
 
       <service-endpoints
+        :endpoints="endpoints"
         class="mb-3"
       />
